+++ include/UnvisitedQueue.h
@@ -15,7 +15,7 @@ public:
   ~UnvisitedQueue() = default;
 
   // Initialize a QueueStop and push it into the internal min-heap using the given priority
-  void push(std::shared_ptr<Stop> stop, float priority);
+  void push(std::shared_ptr<Stop> stop, double priority);
 
   // Clear any processed stops off the top of the internal priority queue. If no 
   // stops remain, return true. If an unprocessed stop remains, return false.
@@ -28,10 +28,10 @@ private:
   // Representation of stops specifically for use in the priority queue
   struct QueueStop {
     std::shared_ptr<Stop> stop;
-    float priority;
+    double priority;
 
     // 2-arg constructor to ensure data members are not left uninitialized
-    QueueStop(std::shared_ptr<Stop> stop, float priority)
+    QueueStop(std::shared_ptr<Stop> stop, double priority)
                  : stop(stop), priority(priority) {}
 
     // Overload < operator for priority queue comparisons
